+++ src/objects/object.rs
@@ -2,7 +2,7 @@ use crate::material::Material;
 use crate::vec3::Vec3;
 
 pub trait Object {
-    fn intersect(&self, orig: Vec3, dir: Vec3) -> (bool, f64);
+    fn intersect(&self, orig: Vec3, dir: Vec3) -> (bool, f32);
 
     fn center(&self) -> Vec3;
