+++ src/raytracing/util.rs
@@ -1,3 +1,3 @@
-pub const EPS: f64 = 0.001;
+pub const EPS: f32 = 0.001;
 pub const MAX_REFLECTION_DEPTH: i32 = 4;
-pub const CLOSEST_VIEW_DISTANCE: f64 = 1e3;
+pub const CLOSEST_VIEW_DISTANCE: f32 = 1e3;
